nearby query + other misc changes

+++ app/src/main/java/com/surrus/galwaybus/model/Location.kt
@@ -1,3 +1,3 @@
 package com.surrus.galwaybus.model
 
-data class Location(val latitude: Float, val longitude: Float)
+data class Location(val latitude: Double, val longitude: Double)
